Editing JUnit tests.
Previous tests did not consider the new attribute (ID) for MenuItem.
Also made the MenuItem tests more efficient, as per Daniel's request.
Worked on by Jonny, Brandon and Matt.
